Move functions out of run.py and into a package in src/selectarr.
