+++ .gitignore
@@ -1,1 +1,2 @@
 venv/
+src/selectarr/__pycache__/
